Exercises

Documentation

readTwoArgs :: [String] -> String

Exercise 1

Change the program so it reads two arguments from the command line, and prints out a message using both of them.

arithmetic :: [String] -> String

Exercise 2

Change the program so it performs a simple arithmetic operation on the two arguments and prints out the result.

promptForName :: IO ()

Exercise 3

getLine is an IO action that reads a line from the console and returns it as a string. Change the program so it prompts for a name, reads the name, and then prints that instead of the command line value.

parseNumber :: Parser LispVal

Exercise 4

Rewrite parseNumber, without liftM, using

do-notation

explicit sequencing with the >>= operator

parseNumber' :: Parser LispVal

Exercise 4

Equivalent version of parseNumber using >>=.

parseEsc :: a

Exercise 5

Our strings aren't quite R5RS compliant, because they don't support escaping of internal quotes within the string. Change parseString so that " gives a literal quote character instead of terminating the string. You may want to replace noneOf """ with a new parser action that accepts either a non-quote character or a backslash followed by a quote mark.